Consolidate classifier and utils
