+++ backend/__init__.py
@@ -2,3 +2,4 @@ from .analyzer import VideoAnalyzer
 from .validator import CampaignValidator
 from .classifier import MilkMobClassifier
 from .tag_detector import CampaignTagDetector
+from .video_utils import save_uploaded_video, extract_thumbnail, process_video_post
